requirements.txt
streamlit==1.37.0
tensorflow==2.15.0
numpy==1.26.4
pillow==10.4.0
opencv-python-headless==4.10.0.84

torch==2.3.1
transformers==4.57.3
safetensors==0.4.5

plotly==5.22.0
streamlit-option-menu==0.3.13
streamlit-extras==0.5.0